update and delete

+++ lib/app/routers/users.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI,HTTPException,Depends,APIRouter,status
 from sqlalchemy.orm import Session
-from .. import schemas,models,database,utils
+from .. import schemas,models,database,utils,oauth2
 
 
 router=APIRouter(prefix="/user",tags=["Users"])
@@ -37,29 +37,35 @@ def create_user(user:schemas.CreateUser,db:Session=Depends(database.get_db)):
 
 
 @router.post("/delete_user/{id}")
-def delete_user(id:int,db:Session=Depends(database.get_db)):
+def delete_user(id:int,db:Session=Depends(database.get_db),current_user:int=Depends(oauth2.get_current_user)):
     
     user=db.query(models.Users).filter(models.Users.uid==id).first()
     if(not user):
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail=f"User with uid {id} does not exist!")
+    
+    if(user.uid!=current_user.uid):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,detail="Not authorized")
+    
+    
     db.delete(user)
     db.commit()
     return {"message":"User deleted successfully"}
 
-@router.put("/update_user/{id}")
-
-
-
-def update_user(id:int,user:schemas.CreateUser,db:Session=Depends(database.get_db)):
-    
+@router.put("/update_user/{id}",status_code=status.HTTP_201_CREATED,response_model=schemas.UserOut)
+def update_user(id:int,user:schemas.CreateUser,db:Session=Depends(database.get_db),current_user:int=Depends(oauth2.get_current_user)):
     
     if(len(str(user.mobile))!=10):
         raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST,detail="Invalid mobile number!")
     
     updated_user=db.query(models.Users).filter(models.Users.uid==id)
+    
     if(not updated_user):
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,detail=f"User with uid {id} does not exist!")
     
+    if(updated_user.first().uid!=current_user.uid):
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,detail="Not authorized")
+    
     updated_user.update(user.model_dump(),synchronize_session=False)
     db.commit()
+    
     return user
